+++ src/client/java/crying/tools/CryingClient.java
@@ -1,7 +1,6 @@
 package crying.tools;
 
 import crying.tools.entities.CrierEntityRenderer;
-import crying.tools.entities.CryingCatEntityRenderer;
 import crying.tools.other.CryingAttribute;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.fabric.api.client.rendering.v1.EntityRendererRegistry;
@@ -33,6 +32,5 @@ public class CryingClient implements ClientModInitializer
         }
 
         EntityRendererRegistry.register(Crying.CRIER, CrierEntityRenderer::new);
-        EntityRendererRegistry.register(Crying.CRYING_CAT, CryingCatEntityRenderer::new);
     }
 }