+++ src/main/java/crying/tools/armors/CryingArmor.java
@@ -59,8 +59,7 @@ public class CryingArmor implements ArmorMaterials {
             Crying.helmet,
             Crying.chestplate,
             Crying.leggings,
-            Crying.boots,
-            Crying.crying_cat_item
+            Crying.boots
         };
         if (itemStack != null) {
             boolean returner = false;